backend-tqf/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.8.2
google-generativeai>=0.7.0
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
